refactor(uikit): Remove react router dom peerDependencies (#371)
* refactor(uikit): Remove react router dom peerDependencies

* test(uikit): Update snapshot

* fix(uikit): Eslint
