+++ packages/pancake-uikit/package.json
@@ -31,6 +31,8 @@
     "@testing-library/jest-dom": "^5.11.6",
     "@testing-library/react": "^11.2.5",
     "@types/react-dom": "^17.0.5",
+    "@types/lodash": "^4.14.168",
+    "@types/styled-system": "^5.1.10",
     "jest-styled-components": "^7.0.3",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
@@ -40,13 +42,10 @@
   "peerDependencies": {
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.0",
     "styled-components": "^5.2.3"
   },
   "dependencies": {
     "@popperjs/core": "^2.9.2",
-    "@types/lodash": "^4.14.168",
-    "@types/styled-system": "^5.1.10",
     "lodash": "^4.17.20",
     "react-popper": "^2.2.5",
     "styled-system": "^5.1.5"
