+++ packages/pancake-uikit/src/components/BaseMenu/index.stories.tsx
@@ -1,4 +1,5 @@
 import React, { useState } from "react";
+// eslint-disable-next-line import/no-extraneous-dependencies
 import { BrowserRouter, Link } from "react-router-dom";
 import InlineMenu from "./InlineMenu";
 import SubMenuComp from "./SubMenu";
