+++ packages/pancake-uikit/src/components/BottomNavItem/index.stories.tsx
@@ -1,4 +1,5 @@
 import React from "react";
+// eslint-disable-next-line import/no-extraneous-dependencies
 import { BrowserRouter } from "react-router-dom";
 import { Flex } from "../Box";
 import BottomNavItem from "./BottomNavItem";
