+++ packages/pancake-uikit/src/components/Button/IconButton.tsx
@@ -1,6 +1,7 @@
 import styled from "styled-components";
+import { PolymorphicComponent } from "../../util/polymorphic";
 import Button from "./Button";
-import { BaseButtonProps, PolymorphicComponent } from "./types";
+import { BaseButtonProps } from "./types";
 
 const IconButton: PolymorphicComponent<BaseButtonProps, "button"> = styled(Button)<BaseButtonProps>`
   padding: 0;
