+++ packages/pancake-uikit/src/components/Button/index.stories.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable import/no-extraneous-dependencies */
 import { capitalize } from "lodash";
 import React, { useState } from "react";
 import { BrowserRouter, Link } from "react-router-dom";
