+++ packages/pancake-uikit/src/components/Button/types.ts
@@ -1,6 +1,6 @@
-import { ComponentProps, ElementType, ReactElement, ReactNode } from "react";
-import { Link } from "react-router-dom";
+import { ElementType, ReactNode } from "react";
 import { LayoutProps, SpaceProps } from "styled-system";
+import type { PolymorphicComponentProps } from "../../util/polymorphic";
 
 export const scales = {
   MD: "md",
@@ -22,23 +22,8 @@ export const variants = {
 export type Scale = typeof scales[keyof typeof scales];
 export type Variant = typeof variants[keyof typeof variants];
 
-/**
- * @see https://www.benmvp.com/blog/polymorphic-react-components-typescript/
- */
-export type AsProps<E extends ElementType = ElementType> = {
-  as?: E;
-};
-
-export type MergeProps<E extends ElementType> = AsProps<E> & Omit<ComponentProps<E>, keyof AsProps>;
-
-export type PolymorphicComponentProps<E extends ElementType, P> = P & MergeProps<E>;
-
-export type PolymorphicComponent<P, D extends ElementType = "button"> = <E extends ElementType = D>(
-  props: PolymorphicComponentProps<E, P>
-) => ReactElement | null;
-
 export interface BaseButtonProps extends LayoutProps, SpaceProps {
-  as?: "a" | "button" | typeof Link;
+  as?: "a" | "button" | ElementType;
   external?: boolean;
   isLoading?: boolean;
   scale?: Scale;
