+++ packages/pancake-uikit/src/components/DropdownMenu/index.stories.tsx
@@ -1,4 +1,5 @@
 import React from "react";
+// eslint-disable-next-line import/no-extraneous-dependencies
 import { BrowserRouter } from "react-router-dom";
 import { Box } from "../Box";
 import DropdownMenu from "./DropdownMenu";
