+++ packages/pancake-uikit/src/components/Footer/index.stories.tsx
@@ -1,5 +1,6 @@
 import { noop } from "lodash";
 import React from "react";
+// eslint-disable-next-line import/no-extraneous-dependencies
 import { BrowserRouter } from "react-router-dom";
 import { langs, footerLinks } from "./config";
 import Footer from "./Footer";
