+++ packages/pancake-uikit/src/widgets/Menu/context.ts
@@ -1,0 +1,3 @@
+import { createContext, ElementType } from "react";
+
+export const MenuContext = createContext<{ linkComponent: ElementType }>({ linkComponent: "a" });
